requirements.txt
Flask==3.0.2
numpy==1.24.4
smart_open==6.3.0
opencv_python==4.7.0.68
opencv_python_headless==4.7.0.72
Pillow==9.3.0
Requests==2.31.0
SQLAlchemy
psycopg2-binary
flask-cors
